+++ source/graphics/Image.cpp
@@ -60,8 +60,8 @@ void Image::load(const std::string &textureName) {
 void Image::load(const Texture &texture) {
 	m_texture = &texture;
 
-	m_width = m_texture->getSize().x;
-	m_height = m_texture->getSize().y;
+	m_width = (u16)m_texture->getSize().x;
+	m_height = (u16)m_texture->getSize().y;
 
 	setClipRect(0, 0, m_width, m_height);
 	setPosRect(0, 0, m_width, m_height);
